+++ META-INF/MANIFEST.MF
@@ -1,0 +1,1 @@
+Main-Class: com.fedapp.App
