+++ src/test/java/com/fedapp/AppTest.java
@@ -1,4 +1,4 @@
-package com.fedepot;
+package com.fedapp;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
@@ -7,7 +7,7 @@ import junit.framework.TestSuite;
 /**
  * Unit test for simple App.
  */
-public class AppTest 
+public class AppTest
     extends TestCase
 {
     /**
